pyproject.toml
[project]
name = "featurebench"
version = "0.1.0"
description = "FeatureBench Pipeline - A test-driven data generation pipeline for building and evaluating feature-level coding benchmarks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.3.3",
    "docker>=5.0.3",
    "filelock>=3.20.0",
    "huggingface-hub>=0.26.5",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.7",
    "openai>=2.6.1",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "rich>=13.0.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
]

[project.scripts]
fb = "featurebench.cli:main"
featurebench = "featurebench.cli:main"

[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["featurebench"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
